+++ Kernel/System/SupportDataCollector/Plugin/Database/mysql/Charset.pm
@@ -59,8 +59,9 @@ sub Run {
                 Identifier => 'ServerEncoding',
                 Label      => Translatable('Server Database Charset'),
                 Value      => $Row[1],
-                Message =>
-                    "This character set is not yet supported, please see https://bugs.otrs.org/show_bug.cgi?id=12361. Please convert your database to the character set 'utf8'.",
+                Message    => Translatable(
+                    "This character set is not yet supported, please see https://bugs.otrs.org/show_bug.cgi?id=12361. Please convert your database to the character set 'utf8'."
+                ),
             );
         }
         elsif ( $Row[1] =~ /utf8/i ) {